+++ src/main/java/com/brachii/ktelabstest/service/PatientServiceImpl.java
@@ -1,6 +1,8 @@
 package com.brachii.ktelabstest.service;
 
 import com.brachii.ktelabstest.dto.PatientDto;
+import com.brachii.ktelabstest.exception.EmptyValueException;
+import com.brachii.ktelabstest.exception.NotFoundException;
 import com.brachii.ktelabstest.model.Patient;
 import com.brachii.ktelabstest.repository.PatientRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +24,10 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public Patient createPatient(Patient patient) {
+        if (patient == null) {
+            throw new EmptyValueException("Patient cannot be null");
+        }
+
         UUID uuidNew = UUID.randomUUID();
         patient.setUuidPatient(uuidNew);
         Patient newPatient = patientRepository.save(patient);
@@ -31,10 +37,15 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public PatientDto getPatientById(Long patientId) {
+        if (patientId == null) {
+            throw new EmptyValueException("Id cannot be null");
+        }
+
         Optional<Patient> patientOptional = patientRepository.findById(patientId);
 
         if (patientOptional.isEmpty()) {
             log.info("Patient with id=" + patientId + " not found.");
+            throw new NotFoundException("Patient with id=" + patientId + " not found.");
         }
 
         Patient patient = patientOptional.get();
@@ -45,6 +56,11 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public void deletePatient(Long patientId) {
+        if (patientId == null) {
+            throw new EmptyValueException("Id cannot be null");
+        }
+
+        getPatientById(patientId);
         patientRepository.deleteById(patientId);
         log.info("Patient with id=" + patientId + " deleted.");
     }
